django_backend/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.7
django-allauth==0.54.0
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.0
whitenoise==6.6.0
gunicorn==21.2.0

# Scientific computing libraries for AHP analysis
numpy==1.24.3
scipy==1.11.3
pandas==2.1.1

# Image processing for user avatars
Pillow==10.0.1